pyproject.toml
[project]
name = "kanban-agents"
version = "0.1.0"
description = "Two-agent Kanban loop for software work using Claude CLI"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.12",
    "uvicorn[standard]==0.34.2",
    "pyyaml==6.0.2",
    "pyjwt[crypto]>=2.12.1",
]

[tool.uv]
dev-dependencies = [
    "httpx==0.28.1",
    "pytest==8.3.5",
    "pytest-asyncio==0.25.3",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
